requirements.txt
dotenv
asyncio
datetime
loguru
pipecat-ai==0.0.102
pipecat-ai-tail==0.0.2
pipecat-ai[google]
pipecat-ai[whisper]
pipecat-ai[piper]
google-genai
torch
onnxruntime-gpu
